chore(docs-builder): remove variable statements(include spyOn) from example code
