+++ package.json
@@ -26,7 +26,6 @@
     "build:markdown": "remark README.md README.zh.md docs/cli*.md docs/cookbook*.md docs/agent-development*.md -o"
   },
   "devDependencies": {
-    "@aigne/typedoc-plugin-example-utils-ts": "^1.0.3",
     "@biomejs/biome": "^1.9.4",
     "@biomejs/js-api": "^0.7.1",
     "@biomejs/wasm-nodejs": "^1.9.4",
